Add files via upload
In case anyone was workign on this tonight, I uploaded some bits of information about the emissions data, and included my graph.  It's in the version "Project1ppt_robby"  feel free to add on to it, or merge with what you have.  Otherwise will be in early on Wednesday to go over in person